fix component namespacing issues
