+++ priv/templates/talon.gen.theme/admin-lte/templates/layout/sidebar.html.slim
@@ -1,11 +1,13 @@
 aside.main-sidebar
   section.sidebar style="height: auto;"
     ul.sidebar-menu
-      li.header.text-uppercase Resources
+      li.header= dgettext "talon", "Resources"
+      / TODO: Dont' think we need this if check anymore. Remove
       = if @talon_resource do
-        = for resource_path <- Talon.Concern.resource_paths(@conn) do
+        = for resource_path <- resource_paths(@conn) do
           = render "nav_resource_link.html", conn: @conn, resource_path: resource_path
-      li.header.text-uppercase Actions
+      li.header= dgettext "talon", "Actions"
+      / TODO: Dont' think we need this if check anymore. Remove
       = if @talon_resource do
-        = for action_path <- Talon.Concern.nav_action_links(@conn) do
+        = for action_path <- nav_action_links(@conn) do
           = render "nav_action_link.html", conn: @conn, action_path: action_path
